services/calculation-engine.bak/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calculation-engine"
version = "0.1.0"
description = "EnGHub Calculation Platform Engine - Foundation Phase"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sympy>=1.12",
    "pint>=0.22.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.isort]
profile = "black"
line-length = 100

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "api: API endpoint tests",
]
